the_project/broadcaster/requirements.txt
flask
nats-py
requests
